requirements.txt
openai
colorama
sounddevice
numpy
scipy
keyboard
wave
faster-whisper
pyaudio
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
transformers
pillow
opencv-python
pywin32
TTS
termcolor